pyproject.toml
[project]
name = "intents-dashboard"
version = "0.1.0"
description = "NEAR intents fees dashboard (Streamlit + Snowflake)"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.36.0",
    "pandas>=2.2.0",
    "python-dotenv>=1.0.0",
    "snowflake-connector-python>=3.7.0",
    "plotly>=5.22.0",
    "cryptography>=42.0.0"
]
